refactor and let the old api still work
